fix: wrong HTTP header and parse method
1. When sending a HTTP request, set "Connection" field to "close"
would close the connection, so we need to re-establish the connection
between our machine and proxy to send another reuqests. So, we
need to specify it to be "keep-alive".
2. Since setting "Connection" to "keep-alive", SSL_write() can not
recognize the end of request, so we need to parse response header to
get the "Content-Length" field to get the length of the response body
and determine the end of HTTP response manually.